NLP_Response/requirements.txt
fastapi
uvicorn
joblib
numpy
google-cloud-storage
scikit-learn
spacy